Call API
the callAPI function calls the zelda compendium API a set number of times based on the difficulty setting. Each call gets a random entry by utilizing the math.random built in js function. Once fetch attempt has been succesful the image url is extrapulated from the data. Then, the image url is pushed to an array.

When all the fetch requests have been completed, the entry usestate is updated with the array.