Round logic
Moved logic from Start component to current file to make round logic possible.

The round logic is made from a useEffect hook that checks the current points for either multiples of 3 or 8 depending on the difficulty. This is achieved by taking the current points and checking the remaineder, if the remaineder is equal to zero then the next round is initiated with a new array of images. If not, the function is returned.